added base atlas class
